fix build on windows

+++ src/support/allocators/zeroafterfree.h
@@ -37,7 +37,7 @@ struct zero_after_free_allocator : public std::allocator<T> {
     void deallocate(T* p, std::size_t n)
     {
         if (p != nullptr)
-            memory_cleanse(p, sizeof(T) * n);
+            memory_cleanse(p, (size_t) sizeof(T) * n);
         std::allocator<T>::deallocate(p, n);
     }
 };
